wpt: refactoring and docs
